requirements.txt
streamlit==1.34.0
pandas==2.2.2
gunicorn==22.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1
firebase_admin==6.5.0